Fix crash when using user-defined files without a debugger
On some systems, when no debuggers are attached to the program, when
clicking on "Install for ME1" after supplying at least one
user-supplied file, a Windows crash dialog is shown. The dialog does not
close the program however.

This is due to calling Debugger.Break() even though no debuggers are
attached to the program, which shows a crash dialog on some systems.

This adds a check before using Debugger.Break() (in the two only
places where it is used) for whether a debugger is currently attached.
This both keeps the behaviour of interrupting if a debugger is attached,
and prevents showing a crash dialog if no debuggers are attached.

See: https://xamarin.github.io/bugzilla-archives/33/33952/bug.html